requirements.txt
fastapi==0.135.1
requests==2.32.3
bs4==0.0.2
playwright==1.57.0
langchain==0.3.24
langchain_community==0.3.22
uvicorn==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
langchain==0.3.24
langchain-chroma==0.1.4
langchain-community==0.3.22
langchain-core==0.3.55
langchain-experimental==0.3.4
langchain-google-genai==2.0.3
langchain-openai==0.2.11
langchain-text-splitters==0.3.8